kb_desktop/requirements.txt
PySide6
python-docx
faiss-cpu
numpy
# openai # Uncomment when needed, or install now
openai
python-dotenv
pyinstaller
